# docs/API_SPEC.yaml
openapi: 3.0.3
info:
  title: Prospect API
  version: 0.1.0
  description: Evidence-first financial document intelligence API.
servers:
  - url: /api/v1
security:
  - sessionToken: []

# Every non-2xx response body is an Error. A document owned by another user returns the
# same 404 as a nonexistent one.
paths:
...
  /documents:
    post:
      summary: Create a document and a signed upload URL (direct-to-storage upload)
      description: >
        The browser PUTs the PDF to `upload.url` with `upload.headers`, then calls
        POST /documents/{documentId}/complete. The URL expires at `upload.expires_at`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentCreate"
      responses:
        "201":
          description: Document created in UPLOADING state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentUpload"
        "413":
          description: Declared size exceeds MAX_UPLOAD_BYTES (code file_too_large)
        "415":
          description: content_type is not application/pdf (code unsupported_file_type)

    get:
      summary: List the current user's documents, newest first
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: Document list
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Document"
  /documents/{documentId}/complete:
    post:
      summary: Verify the uploaded bytes server-side and queue processing
      description: >
        Checks the stored object's size and PDF signature. Success sets UPLOADED and creates a
        processing job. Idempotent once the document has left UPLOADING.
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "200":
          description: Document (UPLOADED)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "409":
          description: File not uploaded yet (code upload_not_found); safe to retry
        "422":
          description: Stored file is not a valid PDF or its size differs (code invalid_upload); document becomes FAILED
  /documents/{documentId}/download-url:
    get:
      summary: Short-lived signed URL to read the private PDF
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "200":
          description: Signed URL
          content:
            application/json:
              schema:
                type: object
                required: [url, expires_at]
                properties:
                  url:
                    type: string
                  expires_at:
                    type: string
                    format: date-time
        "409":
          description: File not available (UPLOADING or FAILED)
  /documents/{documentId}:
    get:
      summary: Get document metadata and processing state
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"

    patch:
      summary: Update document metadata (company name)
      description: >
        The company name groups the user's reports into one company workspace (whitespace is
        collapsed; matching is case-insensitive). null clears it.
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [company_name]
              properties:
                company_name:
                  type: string
                  maxLength: 200
                  nullable: true
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"

    delete:
      summary: Delete a document
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "204":
          description: Deleted
  /documents/{documentId}/pages:
    get:
      summary: Page summaries in page order (available once the document is processed)
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "200":
          description: Pages
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      required: [page_number, extraction_status, char_count]
                      properties:
                        page_number:
                          type: integer
                          description: 1-based physical page, as a PDF viewer numbers it
                        label:
                          type: string
                          nullable: true
                          description: Printed page label defined by the PDF, if any
                        extraction_status:
                          type: string
                          enum: [success, partial, failed]
                        char_count:
                          type: integer
  /documents/{documentId}/sections:
    get:
      summary: Detected sections in document order
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "200":
          description: Sections
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      required: [id, ordinal, start_page, end_page]
                      properties:
                        id:
                          type: string
                          format: uuid
                        ordinal:
                          type: integer
                        title:
                          type: string
                          nullable: true
                          description: Verbatim heading text; null when no heading was detected
                        start_page:
                          type: integer
                        end_page:
                          type: integer
  /documents/{documentId}/pages/{pageNumber}:
    get:
      summary: Get one page's extracted text
      parameters:
        - $ref: "#/components/parameters/DocumentId"
        - name: pageNumber
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Page
          content:
            application/json:
              schema:
                type: object
                required: [page_number, extraction_status, text]
                properties:
                  page_number:
                    type: integer
                  label:
                    type: string
                    nullable: true
                  extraction_status:
                    type: string
                    enum: [success, partial, failed]
                  text:
                    type: string
                  width:
                    type: number
                    nullable: true
                  height:
                    type: number
                    nullable: true
        "404":
          description: Document or page not found
...
  /documents/{documentId}/calculations:
    get:
      summary: Get deterministic ratios with their formula and source facts
      description: >
        Computed in application code from accepted facts, never by an LLM. value is a plain ratio
        at full precision as a decimal string (0.181 = 18.1%); round only for display. A missing
        input, a zero denominator, or incompatible inputs give status not_possible with a reason,
        never a zero.
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "200":
          description: Calculations
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Calculation"
...
  /companies/{companyId}/documents:
    post:
      summary: Associate a document with a company
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyDocumentAttach"
      responses:
        "200":
          description: Updated company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyOut"
  /companies/{companyId}/documents/{documentId}:
    delete:
      summary: Detach a document from a company
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "204":
          description: Document detached
...
components:
  securitySchemes:
    sessionToken:
      type: http
      scheme: bearer
  parameters:
    DocumentId:
      name: documentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    Calculation:
      type: object
      required: [id, metric, name, formula_key, formula, period_type, period_label, status, value, unit, notes, inputs]
      properties:
        id:
          type: string
          format: uuid
        metric:
          type: string
          enum: [revenue_growth, net_margin, roa, roe, debt_to_equity, current_ratio]
        name:
          type: string
        formula_key:
          type: string
          enum: [revenue_growth, net_margin, roa_average_assets, roa_ending_assets, roe_average_equity, roe_ending_equity, debt_to_equity, current_ratio]
          description: The *_ending_* variants are used, and labeled, only when the opening balance is missing
        formula:
          type: string
        period_type:
          type: string
          enum: [annual, instant]
        period_label:
          type: string
          description: FY2025 for annual ratios, the balance-sheet date for debt-to-equity and current ratio
        status:
          type: string
          enum: [calculated, not_possible]
        value:
          type: string
          nullable: true
          description: Plain ratio, full precision; null unless calculated
        unit:
          type: string
          enum: [percent, times]
          description: Presentation only
        reason_code:
          type: string
          nullable: true
          enum: [MISSING_INPUT, DIVISION_BY_ZERO, INCOMPATIBLE_INPUTS, null]
        reason:
          type: string
          nullable: true
          description: e.g. "Total equity at 2025-12-31 was not found."
        notes:
          type: array
          items:
            type: string
          description: Caveats, e.g. a negative denominator or the ending-balance variant
        inputs:
          type: array
          description: Source facts (with evidence) used, or those found when not possible
          items:
            $ref: "#/components/schemas/FinancialFact"
...
    DocumentCreate:
      type: object
      required: [filename, content_type, size_bytes]
      properties:
        filename:
          type: string
          maxLength: 255
        content_type:
          type: string
          enum: [application/pdf]
        size_bytes:
          type: integer
          minimum: 1
        document_type:
          type: string
          enum: [annual_report, financial_statement, prospectus]
          default: annual_report
        fiscal_year:
          type: integer
          nullable: true
        company_name:
          type: string
          maxLength: 200
          nullable: true
          description: Groups the user's reports into one company workspace
        sha256:
          type: string
          pattern: "^[0-9a-f]{64}$"
          description: SHA-256 of the file; the worker refuses stored bytes that do not match
    DocumentUpload:
      type: object
      required: [document, upload]
      properties:
        document:
          $ref: "#/components/schemas/Document"
        upload:
          type: object
          required: [url, method, headers, expires_at]
          properties:
            url:
              type: string
            method:
              type: string
              enum: [PUT]
            headers:
              type: object
              additionalProperties:
                type: string
            expires_at:
              type: string
              format: date-time
    Document:
      type: object
      required: [id, filename, document_type, mime_type, size_bytes, status, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
        document_type:
          type: string
          enum: [annual_report, financial_statement, prospectus]
        fiscal_year:
          type: integer
          nullable: true
        company_name:
          type: string
          nullable: true
        company_id:
          type: string
          format: uuid
          nullable: true
        mime_type:
          type: string
        size_bytes:
          type: integer
        status:
          type: string
          enum: [UPLOADING, UPLOADED, QUEUED, PROCESSING, EXTRACTING, INDEXING, READY, FAILED]
        processing_error:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    CompanyOut:
      type: object
      required: [id, user_id, name, currency, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
        name:
          type: string
        ticker:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        currency:
          type: string
        description:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        documents:
          type: array
          items:
            type: object
        latest_period:
          type: string
          nullable: true
        latest_revenue:
          type: string
          nullable: true
        latest_net_income:
          type: string
          nullable: true
    CompanyDocumentAttach:
      type: object
      required: [document_id]
      properties:
        document_id:
          type: string
          format: uuid
...